requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GPIO and Hardware Control
RPi.GPIO==0.7.1
spidev==3.6
smbus2==0.4.3

# Serial Communication for Barcode Reader
pyserial==3.5

# Database Connectivity
aiosqlite==0.19.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# HTTP Client for Remote Database
httpx==0.25.2
aiohttp==3.9.1

# Logging and Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and Environment
python-dotenv==1.0.0
click==8.1.7
PyYAML==6.0.1

# Date/Time handling
python-dateutil==2.8.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
